From Christophe Loustaunau,
First submission: "But with vpbmaster I have a little problem :

If I use the option --run-path d:\tmp
vpbmaster write in the task files : --run-path C:\Users\Christophe 

It seems to use the default path and not the path I have specified.
I have just added the line

taskManager->setRunPath(runPath);

when vpbMaster read the --run-path option and it works. I don't know if it's the better solution but anyway
here is the file modified vpbmaster.cpp"


Second submission: "If osg is compiled with OSG_USE_FLOAT_BOUNDINGSPHERE = OFF
There is compile errors with VPB :

osg::maximum(cutOffDistance, pagedLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio());

cutOffDistance is a float and the other part a double.

Here is the file Destination.cpp modified"

+++ src/vpb/Destination.cpp
@@ -2871,7 +2871,7 @@ osg::Node* CompositeDestination::createScene()
         }
     }
         
-    cutOffDistance = osg::maximum(cutOffDistance, myLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio());
+    cutOffDistance = osg::maximum(cutOffDistance, (float)(myLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio()));
     
     myLOD->setRange(0,cutOffDistance,farDistance);
     myLOD->setRange(1,0,cutOffDistance);
@@ -3191,7 +3191,7 @@ osg::Node* CompositeDestination::createPagedLODScene()
         }
     }
         
-    cutOffDistance = osg::maximum(cutOffDistance, pagedLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio());
+    cutOffDistance = osg::maximum(cutOffDistance,(float)(pagedLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio()));
     
     pagedLOD->setRange(0,cutOffDistance,farDistance);
